refactor: implement template design pattern
