+++ app/src/main/java/com/example/zoomwroom/RiderHomeActivity.java
@@ -3,15 +3,12 @@ package com.example.zoomwroom;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Parcelable;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
-import android.widget.FrameLayout;
 import android.widget.TextView;
 
 import androidx.annotation.Nullable;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 import androidx.fragment.app.FragmentManager;
@@ -19,7 +16,7 @@ import androidx.fragment.app.FragmentTransaction;
 
 import com.example.zoomwroom.Entities.DriveRequest;
 import com.example.zoomwroom.Entities.Driver;
-import com.example.zoomwroom.Entities.Rider;
+import com.example.zoomwroom.Entities.Location;
 import com.example.zoomwroom.database.MyDataBase;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
@@ -30,7 +27,6 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.vision.Frame;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
@@ -42,8 +38,6 @@ import com.google.firebase.iid.FirebaseInstanceId;
 import com.google.firebase.iid.InstanceIdResult;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.util.ArrayList;
 
 /**
@@ -54,7 +48,7 @@ import java.util.ArrayList;
  * Allows the user to create a ride and see their profile from fragments
  *
  *
- * @see com.example.zoomwroom.Location
+ * @see Location
  * @see com.google.android.gms.maps.GoogleMap
  *
  * March 12, 2020
@@ -62,17 +56,14 @@ import java.util.ArrayList;
  * Modified source from: https://developers.google.com/maps/documentation/android-sdk/start
  * Under the Apache 2.0 license
  */
-public class RiderHomeActivity extends FragmentActivity implements OnMapReadyCallback, GoogleMap.OnMapClickListener, Serializable {
+public class RiderHomeActivity extends MapsActivity implements Serializable {
 
-    private GoogleMap mMap;
-    protected Location mLocation;
     private boolean f;
-    private Marker departureMarker;
-    private Marker destinationMarker;
+
     private FloatingActionButton profileButton;
     private Button rideButton;
     private String riderEmail;
-    private String token;
+
     private TextView rideStatus;
 
 
@@ -82,15 +73,6 @@ public class RiderHomeActivity extends FragmentActivity implements OnMapReadyCal
         setContentView(R.layout.activity_rider_home);
 
 
-
-
-        FirebaseInstanceId.getInstance().getInstanceId().addOnSuccessListener(new OnSuccessListener<InstanceIdResult>() {
-            @Override
-            public void onSuccess(InstanceIdResult instanceIdResult) {
-                token = instanceIdResult.getToken();
-                Log.d("newToken-----", token);
-            }
-        });
 
         //get current user;
         final FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
@@ -105,7 +87,7 @@ public class RiderHomeActivity extends FragmentActivity implements OnMapReadyCal
                         ArrayList<DriveRequest> requests =  MyDataBase.getRiderRequest(riderEmail);
                         for (DriveRequest request :requests) {
                             if (request.getStatus() == 1) {
-                                //Log.d("newToken", token);
+                                Log.d("newToken", token);
                                 Driver driver = MyDataBase.getDriver(request.getDriverID());
                                 if (driver != null) {
                                     String name = driver.getName();
@@ -128,12 +110,11 @@ public class RiderHomeActivity extends FragmentActivity implements OnMapReadyCal
                     }
                 });
 
-        mLocation = new Location();
+
         // Obtain the SupportMapFragment and get notified when the map is ready to be used.
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
                 .findFragmentById(R.id.map);
         mapFragment.getMapAsync(this);
-
 
         // User profile button
         profileButton = findViewById(R.id.rider_profile_button);
@@ -161,62 +142,31 @@ public class RiderHomeActivity extends FragmentActivity implements OnMapReadyCal
     }
 
 
-    /**
-     * Manipulates the map once available.
-     * This callback is triggered when the map is ready to be used.
-     * This is where we can add markers or lines, add listeners or move the camera.
-     * If Google Play services is not installed on the device, the user will be prompted to install
-     * it inside the SupportMapFragment. This method will only be triggered once the user has
-     * installed Google Play services and returned to the app.
-     */
     @Override
     public void onMapReady(GoogleMap googleMap) {
-        mMap = googleMap;
-
-        LatLng edmonton = new LatLng(53.5232, -113.5263);
-        mMap.moveCamera(CameraUpdateFactory.zoomTo(15.0f));
-        mMap.moveCamera(CameraUpdateFactory.newLatLng(edmonton));
-
-        mMap.setOnMapClickListener(this);
+        super.onMapReady(googleMap);
         f = true; // Set departure flag as true
-
-        departureMarker = mMap.addMarker(new MarkerOptions()
-                .position(edmonton)
-                .title("Departure")
-                .icon(BitmapDescriptorFactory.defaultMarker(244))
-                .alpha(0.71f));
-        destinationMarker = mMap.addMarker(new MarkerOptions()
-                .position(edmonton)
-                .title("Destination")
-                .icon(BitmapDescriptorFactory.defaultMarker(244))
-                .alpha(0.71f));
-
-        departureMarker.setVisible(false);
-        destinationMarker.setVisible(false);
-
     }
-
     /**
      * Moves the destination and departure markers when user clicks on a point on the map
      * flag f - true = set departure
-     *          false = set destination
+     * false = set destination
+     *
      * @param point
      */
     @Override
     public void onMapClick(LatLng point) {
-        if(f){
+        Log.d("POINT", point.toString());
+        if (f) {
             mLocation.setDepart(point);
             f = false;
-            Log.d("Lon", Double.toString(mLocation.getDepart().longitude));
-            Log.d("Lat", Double.toString(mLocation.getDepart().latitude));
         } else {
             mLocation.setDestination(point);
             f = true;
-            Log.d("Lon", Double.toString(mLocation.getDestination().longitude));
-            Log.d("Lat", Double.toString(mLocation.getDestination().latitude));
         }
         updateMarkers();
     }
+
     /**
      * Function is called when the drive request has been accepted
      * @param fragment
@@ -255,12 +205,12 @@ public class RiderHomeActivity extends FragmentActivity implements OnMapReadyCal
         Log.d("Lat", Double.toString(mLocation.getDepart().latitude));
         Log.d("Lon", Double.toString(mLocation.getDestination().longitude));
         Log.d("Lat", Double.toString(mLocation.getDestination().latitude));
-        Log.d("price", Double.toString(getPrice(5.00, 0.5)));
+        Log.d("price", Double.toString(FareCalculation.getPrice(5.00, 0.5, departureMarker, destinationMarker)));
         b.putDouble("desLat", mLocation.getDestination().latitude);
         b.putDouble("desLon", mLocation.getDestination().longitude);
         b.putDouble("depLat", mLocation.getDepart().latitude);
         b.putDouble("depLon", mLocation.getDepart().longitude);
-        b.putDouble("price", getPrice(5.00, 0.5));
+        b.putDouble("price", FareCalculation.getPrice(5.00, 0.5, departureMarker, destinationMarker));
         b.putString("userID", riderEmail);
 
         FragmentManager fragmentManager = getSupportFragmentManager();
@@ -288,85 +238,6 @@ public class RiderHomeActivity extends FragmentActivity implements OnMapReadyCal
         }
     }
 
-        /**
-         * Moves markers to the current latlng position and updates the estimated fare
-         */
-    public void updateMarkers() {
-        LatLng depart = mLocation.getDepart();
-        LatLng destination = mLocation.getDestination();
-        if(!depart.equals(departureMarker.getPosition())) {
-            departureMarker.setPosition(depart);
-            departureMarker.setVisible(true);
-        }
-        if(!destination.equals(destinationMarker.getPosition())) {
-            destinationMarker.setPosition(destination);
-            destinationMarker.setVisible(true);
-        }
-        double price = getPrice(5.00, 0.5);
-        Log.d("Price", Double.toString(price));
-    }
-
-    /**
-     * gets the recommended fare price using formula baseprice + multiplier * distance between points
-     *
-     * @param basePrice
-     * @param multiplier
-     * @return price
-     */
-    public double getPrice(double basePrice, double multiplier) {
-        if(departureMarker.isVisible() && destinationMarker.isVisible()) {
-            double price = basePrice + multiplier * getDistance();
-            return round(price, 2);
-        } else {
-            return 0;
-        }
-    }
-
-    /**
-     * gets distance in kilometers from two latitude/longitude points
-     * by using the haversine formula : https://www.movable-type.co.uk/scripts/latlong.html
-     *
-     * Adapted from javascript code
-     * @return distance
-     */
-    public double getDistance() {
-        final int R = 6371; // Radius of the earth in Km
-        Double latDistance = toRad(departureMarker.getPosition().latitude
-                - destinationMarker.getPosition().latitude);
-        Double lonDistance = toRad(departureMarker.getPosition().longitude
-                - destinationMarker.getPosition().longitude);
-        Double a = Math.sin(latDistance / 2) * Math.sin(latDistance / 2) +
-                Math.cos(toRad(departureMarker.getPosition().latitude)) *
-                        Math.cos(toRad(destinationMarker.getPosition().latitude)) *
-                        Math.sin(lonDistance / 2) * Math.sin(lonDistance / 2);
-        Double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
-        return R * c;
-    }
-
-    /**
-     * Rounds a double value to int places
-     * Source: https://stackoverflow.com/questions/2808535/round-a-double-to-2-decimal-places
-     * @param value
-     * @param places
-     * @return roundedNum
-     */
-    public static double round(double value, int places) {
-        if (places < 0) throw new IllegalArgumentException();
-
-        BigDecimal bd = BigDecimal.valueOf(value);
-        bd = bd.setScale(places, RoundingMode.HALF_UP);
-        return bd.doubleValue();
-    }
-
-    /**
-     * Converts a degree to a radian value
-     * @param value
-     * @return radian
-     */
-    private static Double toRad(Double value) {
-        return value * Math.PI / 180;
-    }
-
 
     /**
      * Called when user wants to create a ride
@@ -379,12 +250,12 @@ public class RiderHomeActivity extends FragmentActivity implements OnMapReadyCal
         Log.d("Lat", Double.toString(mLocation.getDepart().latitude));
         Log.d("Lon", Double.toString(mLocation.getDestination().longitude));
         Log.d("Lat", Double.toString(mLocation.getDestination().latitude));
-        Log.d("price", Double.toString(getPrice(5.00, 0.5)));
+        Log.d("price", Double.toString(FareCalculation.getPrice(5.00, 0.5, departureMarker, destinationMarker)));
         b.putDouble("desLat", mLocation.getDestination().latitude);
         b.putDouble("desLon", mLocation.getDestination().longitude);
         b.putDouble("depLat", mLocation.getDepart().latitude);
         b.putDouble("depLon", mLocation.getDepart().longitude);
-        b.putDouble("price", getPrice(5.00, 0.5));
+        b.putDouble("price", FareCalculation.getPrice(5.00, 0.5, departureMarker, destinationMarker));
         b.putString("userID", riderEmail);
 
         FragmentManager fragmentManager = getSupportFragmentManager();
